requirements.txt
astroid==2.2.5
autopep8==1.4.4
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
google-api-core==1.9.0
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-firestore==0.32.1
googleapis-common-protos==1.5.9
grpcio==1.20.0
idna==2.8
isort==4.3.17
lazy-object-proxy==1.3.1
mccabe==0.6.1
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.5.0
pylint==2.3.1
pytz==2019.1
requests==2.21.0
rsa==4.0
six==1.12.0
typed-ast==1.3.4
urllib3==1.24.2
wrapt==1.11.1
